+++ frontend/src/agents/SharedAgent.tsx
@@ -6,8 +6,8 @@ import { useParams } from 'react-router-dom';
 import userService from '../api/services/userService';
 import NoFilesDarkIcon from '../assets/no-files-dark.svg';
 import NoFilesIcon from '../assets/no-files.svg';
-import AgentImage from '../components/AgentImage';
 import MessageInput from '../components/MessageInput';
+import { Avatar } from '../components/ui/avatar';
 import Spinner from '../components/Spinner';
 import ConversationMessages from '../conversation/ConversationMessages';
 import { Query } from '../conversation/conversationModels';
@@ -152,10 +152,10 @@ export default function SharedAgent() {
   return (
     <div className="relative h-full w-full">
       <div className="absolute top-5 left-4 hidden items-center gap-3 sm:flex">
-        <AgentImage
+        <Avatar
           src={sharedAgent.image}
           alt="agent-logo"
-          className="h-6 w-6 rounded-full object-contain"
+          imgClassName="h-6 w-6 rounded-full object-contain"
         />
         <h2 className="text-foreground text-lg font-semibold">
           {sharedAgent.name}
